feat: implement reload functionality for TimeSeriesChart and update related configurations
