+++ ui/web/src/modules/Interactive/index.ts
@@ -5,5 +5,4 @@ export * from './PivotTable';
 export * from './SortableList';
 export * from './Switch';
 export * from './TableView';
-export * from './TimeSeriesChartView';
 export * from './Toast';
